requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
numpy>=1.24.0
scipy>=1.11.0
pydantic>=2.5.0
pyyaml>=6.0
websockets>=12.0
